pyproject.toml
[project]
name = "opengs-maptool"
version = "3.0" # Syncronize with config.py
description = "A tool for creating province maps and related files"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{name = "Thomas-Holtvedt", email = "thomas.holtvedt@gmail.com"}]

dependencies = [
    "numpy==2.3.5",
    "pillow==12.0.0",
    "PyQt6==6.10.0",
    "PyQt6-Qt6==6.10.0",
    "PyQt6_sip==13.10.2",
    "scipy==1.16.3",
    "tkinterdnd2==0.4.3",
    "tqdm==4.67.2",
]

classifiers = [
    "Development Status :: 3 - Alpha", # assumed alpha but can be changed
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/Thomas-Holtvedt/opengs-maptool/"
Repository = "https://github.com/Thomas-Holtvedt/opengs-maptool/"
Issues = "https://github.com/Thomas-Holtvedt/opengs-maptool/issues/"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
opengs_maptool = ["**/*"]

[tool.setuptools]
include-package-data = true

[project.gui-scripts]
opengs-maptool-main-gui = "opengs_maptool.main:main_gui"
[project.scripts]
opengs-maptool-main-terminal = "opengs_maptool.main:main_terminal"
